Reorganise the loader directory
