+++ src/ensembl/src/shared/components/help-article/VideoArticle.tsx
@@ -18,7 +18,7 @@ import React, { useState, useEffect, useRef } from 'react';
 
 import useResizeObserver from 'src/shared/hooks/useResizeObserver';
 
-import { CircleLoader } from 'src/shared/components/loader/Loader';
+import { CircleLoader } from 'ensemblRoot/src/shared/components/loader';
 
 import { LoadingState } from 'src/shared/types/loading-state';
 import { VideoArticleData } from 'src/shared/types/help-and-docs/article';
