+++ src/ensembl/src/shared/components/loader/CircleLoader.scss
@@ -1,0 +1,17 @@
+@import 'src/styles/common';
+
+.circleLoader {
+  display: inline-block;
+  border: 3px solid $grey;
+  border-top-color: $red;
+  width: 40px;
+  height: 40px;
+  border-radius: 100%;
+  animation: loader-spin 1.3s linear infinite;
+}
+
+@keyframes loader-spin {
+  to {
+    transform: rotate(360deg);
+  }
+}
